Refactored useLikes, Favorites, MyPosts, PostDetails
